+++ src/main/kotlin/no/nav/syfo/altinnmottak/kafka/domain/KAltinnOppfolgingsplan.kt
@@ -5,5 +5,5 @@ data class KAltinnOppfolgingsplan(
     val fodselsnummer: String,
     val virksomhetsnummer: String,
     val behovForBistandFraNav: Boolean,
-    val opprettet: Long
+    val opprettet: Int
 )
